Use optimized build
If we're doing this in Rust because we need speed, then why not use all the speed we can get?

This changes the instructions to use an optimized build, rather than a debug build. It compiles slower, but will run faster, and I think that's what we want for intensive things like this.